+++ docs/api/categories.json
@@ -2,7 +2,7 @@
   {
     "id": "logs",
     "title": "Logs",
-    "count": 1
+    "count": 2
   },
   {
     "id": "metrics",
