+++ docs/api/search-kibana652.json
@@ -1,5 +1,14 @@
 [
   {
+    "description": "Tetss if no pipeline is set, it defaults to the default one",
+    "download": "/epr/default-pipeline/default-pipeline-0.0.2.tar.gz",
+    "name": "default-pipeline",
+    "path": "/package/default-pipeline-0.0.2",
+    "title": "Default pipeline Integration",
+    "type": "integration",
+    "version": "0.0.2"
+  },
+  {
     "description": "This is the example integration.",
     "download": "/epr/example/example-0.0.2.tar.gz",
     "name": "example",
