+++ testdata/package/default-pipeline-0.0.2/dataset/foo/elasticsearch/ingest-pipeline/default.json
@@ -1,0 +1,4 @@
+{
+  "description": "Pipeline for normalizing envoyproxy logs",
+  "processors": []
+}
