+++ testdata/package/default-pipeline-0.0.2/dataset/foo/manifest.yml
@@ -1,0 +1,4 @@
+title: Foo
+
+# Needs to describe the type of this input
+type: logs
