+++ testdata/package/default-pipeline-0.0.2/manifest.yml
@@ -1,0 +1,8 @@
+format_version: 1.0.0
+
+name: default-pipeline
+description: Tests if no pipeline is set, it defaults to the default one
+version: 0.0.2
+title: Default pipeline Integration
+categories: ["logs"]
+type: integration
