+++ testdata/public/package/example-1.0.0/index.json
@@ -52,7 +52,7 @@
     {
       "title": "Foo",
       "name": "foo",
-      "release": "",
+      "release": "beta",
       "type": "logs",
       "ingest_pipeline": "pipeline-entry"
     }
